pdev:set

+++ ql.sh
@@ -10,18 +10,13 @@ export ql_orange='\033[1;33m'
 export ql_green='\033[1;32m'
 export ql_no_color='\033[0m'
 
+
+
 mkdir -p "$HOME/.quicklock"
 
 # if [[ ! -p "$HOME/.quicklock/ql_named_pipe" ]]; then
 #    mkfifo "$HOME/.quicklock/ql_named_pipe";
 # fi
-
-
-ql_log_colors (){
-    echo "sourcing quicklock.sh"
-    echo -e "${ql_cyan}sourcing quicklock.sh${ql_no_color}";
-    echo "${ql_cyan}sourcing quicklock.sh${ql_no_color}";
-}
 
 ql_add_color(){
     export ql_gray='\033[1;30m'
@@ -37,7 +32,7 @@ ql_print_version (){
 }
 
 ql_echo_current_lockname (){
-   echo "$quicklock_name"
+   echo $(ql_get_lockname)
 }
 
  on_ql_trap (){
@@ -47,13 +42,13 @@ ql_echo_current_lockname (){
 }
 
 ql_unlock_process(){
-#  kill -SIGUSR1 "$1"  #  kill <pid> given by $1, with SIGUSR1 signal...
+  kill -SIGUSR1 "$1"  #  kill <pid> given by $1, with SIGUSR1 signal...
 #  kill -SIGUSR2 "$1"
 
 #  kill -10 "$1" $ on linux
 # echo "$1" > "$HOME/.quicklock/ql_named_pipe"
 
-  kill -TERM "$1";
+#  kill -TERM "$1";
 }
 
 
@@ -81,23 +76,35 @@ ql_unlock_process(){
    for i in $(ls "$HOME/.quicklock/locks"); do  echo -e "${ql_cyan}$home/$i${ql_no_color}"; done;
 }
 
- ql_find () {
 
-    local lockname="$quicklock_name";
+ql_write_message () {
+    local lockname="$(ql_get_lockname "$1")";
+    local pid="$(ql_get_lockowner_pid ${lockname})";
+#   local pid="$(ql_get_lockowner_pid | head -n 1)";
 
-    if [[ ! -z "$1" ]]; then
-      lockname="$1";
-    fi
+   if [[  -z "$pid" ]]; then
+       >&2 echo "quicklock: error: no pid could be found.";
+       return 1;
+   fi
+
+   echo "foo" > "${lockname}/${pid}"
+   echo "sent message"
+}
+
+
+ql_get_lockname (){
+
+    local lockname="$1";
 
     if [[ -z "$lockname" ]]; then
-          echo "quicklock: no lockname is available, defaulting to \$PWD as lockname.";
-          lockname="$PWD.lock";
+        >&2 echo "quicklock: no lockname is available, defaulting to \$PWD as lockname.";
+        lockname="$PWD";
     fi
 
     lockname=$(echo "${lockname}" | tr "/" _)
 
     if [[ "$lockname" =~ [^a-zA-Z0-9\-\_] ]]; then
-        echo -e "${ql_magenta}quicklock: lockname has invalid chars - must be alpha-numeric chars only.${ql_no_color}"
+         >&2 echo -e "${ql_magenta}quicklock: lockname has invalid chars - must be alpha-numeric chars only.${ql_no_color}"
         on_ql_conditional_exit
         return 1;
     fi
@@ -107,13 +114,29 @@ ql_unlock_process(){
     fi
 
    if [[ -z "$lockname" ]]; then
-      echo "quicklock: no lockname is available, please pass a valid lockname.";
+      >&2 echo "quicklock: no lockname is available, please pass a valid lockname.";
       return 1;
    fi
 
+   echo "$lockname"
+
+}
+
+
+ ql_get_lockowner_pid () {
+
+    local lockname="$1";
+
+    if [[ -z "$lockname" ]]; then
+      lockname="$(ql_get_lockname)"
+    fi
+
+#    local lockname="$(ql_get_lockname "$1" | head -n 1)"
+
     # echo the contents of the directory, should log PID if exists
+
     ls "$lockname" || {
-       echo "quicklock: it does not appear that the lock exists for lockname: '$lockname'"
+       >&2 echo "quicklock: it does not appear that the lock exists for lockname: '$lockname'"
        return 1;
      }
 }
@@ -126,19 +149,12 @@ ql_on_named_pipe_msg (){
 
   echo "quicklock: received piped message...$1,$2"
   local ql_msg="$1";
-  local ql_current_pid="$2"
 
-  if [[ -z "$1" || -z "$2" ]]; then
-    echo "quicklock: at least one message was empty so this is a no-op.";
-    return 0;
-  fi
-
-  if [[ "$ql_msg" == "$ql_current_pid" ]]; then
-     echo "quicklock: releasing lock because of piped message."
-     ql_release_lock
-     exit 1;
+  echo "quicklock: releasing lock because of piped message."
+   ql_release_lock
+#     exit 1;
 #     trap - EXIT;
-  fi
+
 }
 
 ql_acquire_lock () {
@@ -163,17 +179,29 @@ ql_acquire_lock () {
     return 1;
   }
 
-  mkdir "${qln}/$$";  # add the PID inside the lock dir
-  export quicklock_name="${qln}";  # export the var *only if* above mkdir command succeeds
+  my_named_pipe="${qln}/$$"
+  mkfifo "${my_named_pipe}" &> /dev/null;  # add the PID inside the lock dir
 
   trap on_ql_trap EXIT;
+
+#  trap on_ql_trap USR1;
+#  trap on_ql_trap USR2;
+#   trap on_ql_trap SIGUSR1;
+#  trap on_ql_trap SIGUSR2;
+#  trap on_ql_trap SIGHUP;
+#  trap on_ql_trap HUP;
+
 
   echo "quicklock: process id requesting lock: $$"
   echo "quicklock: parent process id of above pid: $(ps -o ppid= -p $$)";
   echo -e "${ql_green}quicklock: acquired lock with name '${qln}'${ql_no_color}"
 
-    #   my_named_pipe="$HOME/.quicklock/ql_named_pipe"
-    #   while read line; do ql_on_named_pipe_msg "$line" "$$"; done < ${my_named_pipe} &
+  if [[ "$ql_allow_ipc" == "yes" ]]; then
+
+    echo "quicklock: process is reading from named pipe to listen for incoming messages regarding releasing lock.";
+    while read line; do ql_on_named_pipe_msg "$line" "$$"; done < ${my_named_pipe} &
+
+  fi
 
 }
 
@@ -191,10 +219,23 @@ ql_acquire_lock () {
 
 }
 
+ql_release_lock_force(){
+  ql_release_lock "$@" --force
+}
 
- ql_release_lock () {
 
-   if [[ ! -z "$1" ]]; then
+ql_release_lock () {
+
+    local quicklock_name="";
+    local is_force="nope";
+
+    local last="$(echo ${@: -1})"
+    if [[ "${last}" == "--force" ]]; then
+      echo "${ql_magenta}quicklock: warning using --force.${ql_no_color}";
+      is_force="yes"
+    fi
+
+   if [[ ! -z "$1" && "$1" != "--force" ]]; then
         quicklock_name="${1}";
 
    elif [[ -z "${quicklock_name}" ]]; then
@@ -228,17 +269,24 @@ ql_acquire_lock () {
      return 1;
    fi
 
+   local current_pid="$$"
+#   local pid_inside="$(ls "${quicklock_name}" | head -n 1)";
+   local pid_inside="$(ls "${quicklock_name}")";
+
+
+   if [[ "${is_force}" != "yes" && ! -z "${pid_inside}" ]]; then
+     if [[ "$current_pid" != "$pid_inside" ]]; then
+     >&2 echo -e "${ql_magenta}quicklock: without using --force, cannot release lock that was initiated by a different pid.${ql_no_color}";
+     return 1;
+     fi
+   fi
+
+
    rm -r "${quicklock_name}" &> /dev/null &&
    { echo -e "${ql_orange}quicklock: lock with name '${quicklock_name}' was released.${ql_no_color}";  } ||
-   { echo -e "${ql_magenta}quicklock: no lock existed for lockname '${quicklock_name}'.${ql_no_color}"; ql_maybe_fail; }
+   { >&2 echo -e "${ql_magenta}quicklock: no lock existed for lockname '${quicklock_name}'.${ql_no_color}"; ql_maybe_fail; }
 
    trap - EXIT; # clear/unset trap
-#   trap - SIGINT; # clear/unset trap
-#   trap - SIGTERM; # clear/unset trap
-#   trap - INT; # clear/unset trap
-#   trap - TERM; # clear/unset trap
-#   trap - SIGHUP; # clear/unset trap
-#   trap - SIGTRAP; # clear/unset trap
 
 }
 
@@ -276,12 +324,14 @@ export -f on_ql_conditional_exit;
 export -f on_ql_trap;
 export -f ql_acquire_lock;
 export -f ql_release_lock;
-export -f ql_find;
+export -f ql_get_lockowner_pid;
 export -f ql_ls;
 export -f ql_print_version;
 export -f ql_unlock_process;
 export -f ql_on_named_pipe_msg;
 export -f ql_echo_current_lockname;
+export -f ql_get_lockname;
+export -f ql_write_message;
 
 # that's it lulz
